fix(fee): address CodeRabbit review on PR #161
- DB CHECK constraint close_price > 0 (migration + model) to lock the invariant
  against direct/future writes.
- Atomic bulk close: new set_close_price_if_absent (INSERT ... ON CONFLICT (isin)
  DO NOTHING RETURNING) so "Close all" never clobbers a manual price set
  concurrently between the snapshot and the loop.
- Decouple the agent auth guard from market data: new list_open_isins (skips the
  per-ISIN sidecar price fetch) backs _agent_open_isins, so price/clear writes
  do not depend on the sidecar being reachable.
- Keep VirtualFeeRow.price as Decimal end-to-end (no int truncation); HTML
  formats {:,.0f}, Excel keeps full precision via _num.
- Status label uses close-price vs avg-buy (not fee>0) so a zero-fee profit is
  not mislabelled break-even (admin + agent templates).
- Tests updated (guard tests stub list_open_isins; close-all tests stub
  set_close_price_if_absent) + 3 new if-absent service tests. Suite green
  (653 passed).

+++ mgmt_ui/app/services/profit_report.py
@@ -99,7 +99,7 @@ class VirtualFeeRow:
     symbol: str
     open_qty: int
     avg_buy_price: Decimal
-    price: int          # realization price (the saved close price)
+    price: Decimal      # realization price (the saved close price)
     trigger: str        # "close"
     in_loss: bool
     fee: Decimal
@@ -470,7 +470,7 @@ async def build_fee_report(
                 customer_id=cust_id, agent_id=agent_id_of_group,
                 broker=group[0].broker, isin=_isin, symbol=sym,
                 open_qty=rem_qty, avg_buy_price=avg_buy,
-                price=int(realize_price), trigger="close",
+                price=realize_price, trigger="close",
                 in_loss=in_loss, fee=vfee,
                 oldest_buy_date=oldest,
             ))
